refactor(outstanding-clap): reorganize lib.rs into focused modules
Split the monolithic lib.rs (1,357 lines) into focused modules:

- help/: Help rendering (config.rs, data.rs, render.rs, template.txt)
- dispatch.rs: Command dispatch logic
- result.rs: HelpResult enum
- outstanding.rs: Outstanding and OutstandingBuilder types
- handler.rs: Unchanged (already focused)

Public API remains unchanged - all types re-exported from lib.rs.
All 32 unit tests, 17 integration tests, and doc tests pass.
